requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
d==0.2.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
itypes==1.2.0
Jinja2==2.11.2
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
psycopg2==2.8.5
Pygments==2.6.1
pyquery==1.4.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
whitenoise==5.0.1
